requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiomysql==0.1.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
mysql-connector-python==8.0.33
protobuf==3.20.3
pyaes==1.6.1
pyasn1==0.4.8
PyMySQL==1.0.3
python-dotenv==1.0.0
pytz==2023.3
rsa==4.9
SQLAlchemy==2.0.9
Telethon==1.28.5
typing_extensions==4.5.0
yarl==1.8.2
